signals: centralize signals
Move the signals responsible for invalidating the cached config and
triggering meta categorization on config changes into signals.py
for better organization and maintainability.
